+++ app/schemas/product_variant_schema.py
@@ -5,8 +5,8 @@ from pydantic import BaseModel, Field, model_validator, ConfigDict
 
 class ProductVariantCreate(BaseModel):
     sku: str = Field(..., min_length=3, max_length=50)
-    price: float = Field(..., gt=0)
-    discount_price: Optional[float] = Field(None, gt=0)
+    price: int = Field(..., gt=0)
+    discount_price: Optional[int] = Field(None, gt=0)
     stock: int = Field(..., ge=0)
     attributes: Dict[str, str] = Field(default_factory=dict)
 
@@ -36,8 +36,8 @@ class ProductVariantCreate(BaseModel):
 
 class ProductVariantUpdate(BaseModel):
     sku: str = Field(..., min_length=3, max_length=50)
-    price: Optional[float] = Field(None, gt=0)
-    discount_price: Optional[float] = Field(None, gt=0)
+    price: Optional[int] = Field(None, gt=0)
+    discount_price: Optional[int] = Field(None, gt=0)
     stock: Optional[int] = Field(None, ge=0)
     attributes: Optional[Dict[str, str]] = None
 
@@ -49,7 +49,7 @@ class ProductVariantUpdate(BaseModel):
 
 class ProductVariantResponse(BaseModel):
     sku: str = Field(..., min_length=3, max_length=50)
-    price: float = Field(..., gt=0)
-    discount_price: Optional[float] = Field(None, gt=0)
+    price: int = Field(..., gt=0)
+    discount_price: Optional[int] = Field(None, gt=0)
     stock: int = Field(..., ge=0)
     attributes: Dict[str, str] = Field(default_factory=dict)
